requirements.txt
cffi~=1.16.0
requests~=2.31.0
tqdm~=4.66.1
numpy~=1.26.3
tmu~=0.8.2
scikit-learn~=1.3.2
pandas~=2.1.4
scipy~=1.11.4
keras~=2.15.0
matplotlib~=3.8.2
setuptools~=68.2.2
gymnasium~=0.29.1
torch~=2.2.0+cu121
PyYAML~=6.0.1
wandb~=0.16.3